Refactoring classes

+++ app/Services/ApplicationService.php
@@ -9,79 +9,53 @@ use Illuminate\Support\Facades\Auth;
 class ApplicationService
 {
 
-    public function storeImage($data, $application, $applicationUserState, $request)
+    public function storeImage($data, $application, $request)
     {
-
         if ($files = $request->file('image')) {
-
             $fileImage = Str::random(50)  . '-' . $files->getClientOriginalName();
 
             $path = 'images/applications/';
-
             $pathWhitFile = 'images/applications/' . $fileImage;
-
             $count = $application->where('image_src', $pathWhitFile)->count();
 
             //En caso de que ya exista un id unico para esa imgaen, genero uno nuevo
             while ($count != 0) {
-
                 $fileImage = Str::random(50)  . '-' . $files->getClientOriginalName();
-
                 $count = $application->where('image_src', $path . $fileImage)->count();
             }
 
             $files->move($path, $fileImage);
-
             $data['image_src'] = $pathWhitFile;
-
-            return $this->saveApplication($data, $application, $applicationUserState);
         } else {
             //Si no carga ninguna imgaen, se le asigna una por defecto.
             $data['image_src'] = 'images/default-image.jpg';
-            return $this->saveApplication($data, $application, $applicationUserState);
         }
+        return $data;
     }
-
-    public function saveApplication($data, $application, $applicationUserState)
-    {
-        $myNewApp = $application->create($data);
-        $user = User::findOrFail(Auth::id());
-        $applicationUserState->users()->attach($user, ['application_id' => $myNewApp->id, 'state_id' => 4]);
-    }
-
 
     public function updateImage($application, $request)
     {
 
         if ($files = $request->file('image')) {
-
-
             $path = 'images/applications/';
 
-            if (file_exists($application->image_src)) {
-
+            if (file_exists($application->image_src && $application->image_src != 'images/default-image.jpg')) {
                 unlink($application->image_src);
             }
-
 
             $fileImage = Str::random(50)  . '-' . $files->getClientOriginalName();
 
             $fileForUpdate =  $path . $fileImage;
-
             $files->move($path, $fileImage);
 
-            $data = [
-                'price' => $request->price,
-                'image_src' => $fileForUpdate
-            ];
+            $data['price'] = $request->price;
+            $data['image_src'] = $fileForUpdate;
 
             return $data;
         } else {
-            
-            $data = [
-                'price' => $request->price,
-                'image_src' =>  'images/default-image.jpg'
-            ];
+
+            $data['price'] = $request->price;
+            $data['image_src'] = 'images/default-image.jpg';
 
             return $data;
         }
